Renaming plugs to be more clear
